requirements.txt
Django==2.1.4
django-tinymce==2.8.0
django-tinymce4-lite==1.7.5
gunicorn==19.9.0
jsmin==2.2.2
pytz==2019.1